requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai
python-dotenv
pydantic==2.9.2
